+++ source/commands/create-default/scoreboardreference.js
@@ -29,7 +29,7 @@ async function executeSubcommand(interaction, database, runMode, ...[logicLayer,
 	});
 	const isSeasonal = interaction.options.getString("scoreboard-type") == "season";
 	scoreboard.send({
-		embeds: [isSeasonal ? await buildSeasonalScoreboardEmbed(interaction.guild, database) : await buildOverallScoreboardEmbed(interaction.guild, database)]
+		embeds: [isSeasonal ? await buildSeasonalScoreboardEmbed(interaction.guild, database, logicLayer) : await buildOverallScoreboardEmbed(interaction.guild, database, logicLayer)]
 	}).then(message => {
 		company.scoreboardChannelId = scoreboard.id;
 		company.scoreboardMessageId = message.id;